feat(utils): move universal utils to the universal folder
