+++ .changeset/proud-peas-design.md
@@ -1,0 +1,13 @@
+---
+'@modern-js/builder-shared': patch
+'@modern-js/plugin-data-loader': patch
+'@modern-js/runtime': patch
+'@modern-js/app-tools': patch
+'@modern-js/prod-server': patch
+'@modern-js/server': patch
+'@modern-js/utils': patch
+---
+
+feat(utils): move universal utils to the universal folder
+
+feat(utils): 将运行时使用的 utils 移动到 universal 目录
