+++ packages/cli/plugin-data-loader/src/cli/data.ts
@@ -13,7 +13,7 @@
 import {
   UNSAFE_DeferredData as DeferredData,
   AbortedDeferredError,
-} from '@modern-js/utils/remix-router';
+} from '@modern-js/utils/universal/remix-router';
 
 const DEFERRED_VALUE_PLACEHOLDER_PREFIX = '__deferred_promise:';
 export async function parseDeferredReadableStream(
