+++ packages/cli/plugin-data-loader/src/runtime/index.ts
@@ -10,7 +10,7 @@ import {
   ErrorResponse,
   UNSAFE_DEFERRED_SYMBOL as DEFERRED_SYMBOL,
   type UNSAFE_DeferredData as DeferredData,
-} from '@modern-js/utils/remix-router';
+} from '@modern-js/utils/universal/remix-router';
 import { isPlainObject } from '@modern-js/utils/lodash';
 import { CONTENT_TYPE_DEFERRED, LOADER_ID_PARAM } from '../common/constants';
 import { matchEntry, ServerContext } from '../common/utils';
