+++ packages/runtime/plugin-runtime/src/router/runtime/plugin.node.tsx
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { createStaticHandler } from '@modern-js/utils/remix-router';
+import { createStaticHandler } from '@modern-js/utils/universal/remix-router';
 import {
   createStaticRouter,
   StaticRouterProvider,
