+++ packages/runtime/plugin-runtime/src/runtimeContext.ts
@@ -1,5 +1,5 @@
 import { Store } from '@modern-js-reduck/store';
-import type { StaticHandlerContext } from '@modern-js/utils/remix-router';
+import type { StaticHandlerContext } from '@modern-js/utils/universal/remix-router';
 import { createContext } from 'react';
 import { createLoaderManager } from './core/loader/loaderManager';
 import { runtime } from './core/plugin';
