+++ packages/toolkit/utils/src/index.ts
@@ -1,6 +1,5 @@
 export * from './compiled';
 export * from './commands';
-export * from './format';
 export * from './FileSizeReporter';
 export * from './printBuildError';
 export * from './debug';
