+++ packages/toolkit/utils/src/universal/constants.ts
@@ -1,0 +1,9 @@
+/**
+ * Property mounted on window that describes route manifest
+ */
+export const ROUTE_MANIFEST = `_MODERNJS_ROUTE_MANIFEST`;
+
+/**
+ * hmr socket connect path
+ */
+export const HMR_SOCK_PATH = '/webpack-hmr';
