+++ packages/toolkit/utils/src/universal/formatWebpack.ts
